dont fetch twice ma, its alright

+++ fetch.py
@@ -11,11 +11,12 @@ URL = 'http://opendatalabs.org/misc/openoil/subsidiaries_1.csv'
 res = requests.get(URL)
 for row in unicodecsv.DictReader(StringIO(res.content)):
     url = row.get('file_url')
+    fn = os.path.join(DATA_PATH, hashlib.sha1(url).hexdigest())
+    if os.path.exists(fn):
+        continue
+
     cn = row.get('company_name')
     r = requests.get(url)
-
-    fn = os.path.join(DATA_PATH, hashlib.sha1(url).hexdigest())
-    print url, fn
     with open(fn, 'wb') as fh:
         row['content'] = r.content
         json.dump(row, fh)
